Update library + picture
Fixed a bug which casted the double lux in an integer in the library. Updated the picture where the bug was.

+++ ZimmeranalysatorV2/Code/Libraries/Website/Website.cpp
@@ -13,7 +13,7 @@ String Website::addzero(uint8_t date)
 	else return (String(date));
 }
 
-void Website::printWebsite(double tem, double hum, short lux, short co2, uint16_t year, uint8_t mon, uint8_t mday, uint8_t hour, uint8_t minute, uint8_t sec, WiFiClient client)
+void Website::printWebsite(double tem, double hum, double lux, short co2, uint16_t year, uint8_t mon, uint8_t mday, uint8_t hour, uint8_t minute, uint8_t sec, WiFiClient client)
 {
 	client.println("<html>");
 	client.println("<head>");
